requirements.txt
crc8==0.2.0
paho-mqtt==1.6.1
pkg_resources==0.0.0
pyserial==3.5
PyYAML==6.0.1
